samples: matter: Moved finite map to common util and improved it
Moved finite map implementation from bridge_util to common util
to make it available for all samples.

Made the finite map more versatile, by defining the key as
a template of trivial typename.
